forced to int

+++ vesicle_edge_extractor/vesicle_video.py
@@ -69,7 +69,7 @@ class VesicleVideo:
         self.r_vals = np.full((self.frames.shape[0], self.frames.shape[1]), np.nan)
         self.x_vals = np.full((self.frames.shape[0], self.frames.shape[1]), np.nan)
         self.y_vals = np.full((self.frames.shape[0], self.frames.shape[1]), np.nan)
-        self.status = np.zeros(self.frames.shape[0])
+        self.status = np.zeros(self.frames.shape[0]).astype(int)
 
     def extract_edges(self, extractor_func, curvature_threshold=10):
         """
